+++ labs/fibvec/FibVec.h
@@ -15,13 +15,13 @@ class FibVec {
     ~FibVec();
 
     // Member Functions
-    unsigned int capacity() const;
-    unsigned int count() const ;
-    void insert(int value, unsigned int index); //can throw std::out_of_range exception
-    int lookup(unsigned int index) const; //can throw std::out_of_range exception
+    unsigned long int capacity() const;
+    unsigned long int count() const ;
+    void insert(int value, unsigned long int index); //can throw std::out_of_range exception
+    int lookup(unsigned long int index) const; //can throw std::out_of_range exception
     int pop(); //can throw std::underflow_error exception
     void push(int value);
-    int remove(unsigned int index); //can throw std::out_of_range exception
+    int remove(unsigned long int index); //can throw std::out_of_range exception
 };
 
 #endif
